firmware: main.rs split four ways, the beacon planner in proto with a model, one clock, and the vendored radio as a patch
The hardware loop is a struct in hardware.rs whose effect method
replaces the macro that reached into forty locals; the BLE side is
ble.rs, the deep sleep sleep.rs, the receiver's bookkeeping gpsctl.rs.
main.rs is the boot and nothing else.

When the next beacon goes out is midair_proto::beacon::Planner, driven
against the clock and plan the radio exposes, with unit tests for the
regression this planner once had and an exhaustive model over every
phase of every slot against every pattern of gates, frames arriving
and late passes. The radio's hop is no longer an Option: every config
has a plan, so the second scheduler is gone.

Time is 64-bit everywhere the loop keeps it; the wrapping arithmetic
and the due() helper are gone. The node's dedup tables are the proto's.
The serve loop takes commands from one channel, the settings keep the
five durations in one RTC array indexed by the knob table, and a deep
sleep whose park did not finish is counted in RTC RAM, retried once,
reported on the boot line and in the telemetry.

The vendored esp-radio is recorded as a patch against the 0.17.0
release, with a script that regenerates, applies and checks it.

+++ firmware/src/idle.rs
@@ -79,9 +79,9 @@ pub fn entries() -> u64 {
 /// observable from here; it would have to come from the scheduler's context
 /// switch, which esp-rtos does not expose. A rate that answers "yes it
 /// halts" honestly beats a percentage that would have to be invented.
-pub fn rate(before: u64, after: u64, elapsed_ms: u32) -> u32 {
+pub fn rate(before: u64, after: u64, elapsed_ms: u64) -> u32 {
     if elapsed_ms == 0 {
         return 0;
     }
-    ((after.saturating_sub(before) * 1_000) / u64::from(elapsed_ms)) as u32
+    ((after.saturating_sub(before) * 1_000) / elapsed_ms) as u32
 }
